Generate current state of docs
nbdev can automatically generate documentation for us. The current state
is completely shit, but just going to commit the current documentation
so that we have a starting poing to build off of.

Note that you can preview docs locally with `make docs_serve`. Once we
open source the repo, we can host these docs for free with Github pages.